[Chore] modifier 네이밍 변경
