+++ Gathering/Source/Global/Modifier/RoundButtonModifier.swift
@@ -1,5 +1,5 @@
 //
-//  View+.swift
+//  RoundButtonModifier.swift
 //  Gathering
 //
 //  Created by 김성민 on 11/6/24.
@@ -8,7 +8,7 @@
 import SwiftUI
 
 extension View {
-    func asButtonStyle(foregroundColor: Color,
+    func asRoundButton(foregroundColor: Color,
                        backgroundColor: Color
     ) -> some View {
         self
@@ -18,17 +18,5 @@ extension View {
             .frame(height: 44)
             .background(backgroundColor)
             .clipShape(RoundedRectangle(cornerRadius: 8))
-    }
-    
-    func badge(backgroundColor: Color = Design.green,
-               textColor: Color = Design.white
-    ) -> some View {
-        self
-            .font(.caption)
-            .padding(.horizontal, 8)
-            .padding(.vertical, 2)
-            .background(backgroundColor)
-            .foregroundColor(textColor)
-            .clipShape(Capsule())
     }
 }
